+++ src/config.lua
@@ -1,3 +1,15 @@
 return {
 	enabled = true,
+	Ares = {
+		BloodyWoundOrigination = true,
+		StaffHits = 6,
+		AxeHits = 2,
+		DaggerHits = 10,
+		TorchHits = 5,
+		SkullHits = 3,
+		SuitHits = 6,
+		Duration = 999999,
+		BossDuration = 10
+	}
+	
 }